AbstractPermutations

Documentation for AbstractPermutations.

AbstractPermutations.AbstractPermutation

AbstractPermutations.cycles

AbstractPermutations.degree

AbstractPermutations.firstmoved

AbstractPermutations.fixedpoints

AbstractPermutations.inttype

AbstractPermutations.nfixedpoints

AbstractPermutations.parity

AbstractPermutations.perm

AbstractPermutations.permtype

Base.:^

Base.sign

AbstractPermutations.@perm

AbstractPermutations.AbstractPermutation — Type

AbstractPermutation

Abstract type representing permutations of set 1:n for some n.

Mandatory interface

Subtypes APerm <: AbstractPermutation must implement the following functions:

Base.:^(i::Integer, σ::APerm) - the image of i under σ,

degree(σ::APerm) - the minimal n such that k^σ == k for all k > n,

For primitive ("bare-metal"/"parent-less") permutations one needs to implement

APerm(images::AbstractVector{<:Integer}[, check::Bool=true]) - construct a APerm from a vector of images. Optionally the second argument check may be set to false when the caller knows that images constitute a honest permutation.

Note

There is no formal requirement that the APerm(images) constructor actually returns a APerm. Any AbstractPermutation object would do. This may be useful if constructing permutation from images is not technically feasible.

Note

If APerm is not constructable from type one needs to implement one(::APerm).

Warn

Even though AbstractPermutation <: GroupsCore.GroupElement they don't necessarily implement the whole of GroupElement interface, e.g. it is possible to implement parent-less permutations.

Optional interface

inttype(::Type{<:APerm}) = UInt32 - return the underlying "storage" integer,

if that makes any sense for APerm.

perm(σ::APerm) = σ - return the "bare-metal" permutation (unwrap).

AbstractPermutations.cycles — Method

cycles(g::AbstractPermutation)

Return an iterator over cycles in the disjoint cycle decomposition of g.

AbstractPermutations.degree — Method

degree(σ::AbstractPermutation)

Return a minimal number n such that σ(k) == k for all k > n.

Such number n can be understood as a degree of a permutation, since we can regard σ as an element of Sym(n) (and not of Sym(n-1)).

By convention degree of the trivial permutation must return 1.

AbstractPermutations.firstmoved — Function

firstmoved(g::AbstractPermutation[, range = 1:degree(g)])

Return the first point from range that is moved by g, or nothing if g fixes range point-wise.

AbstractPermutations.fixedpoints — Function

fixedpoints(g::AbstractPermutation[, range = 1:degree(g)])

Return the vector of points in range fixed by g.

AbstractPermutations.inttype — Method

inttype(σ::Type{<:AbstractPermutation})

Return the underlying "storage" integer. For internal use only.

The default is UInt32.

AbstractPermutations.nfixedpoints — Function

nfixedpoints(g::AbstractPermutation[, range = 1:degree(g)])

Return the number of points in range fixed by g.

AbstractPermutations.parity — Method

parity(g::AbstractPermutation)

Return the parity of number of factors in factorization of g into transpositions.

Return 1 if the number is odd and 0 otherwise.

AbstractPermutations.perm — Method

perm(p::AbstractPermutation)

Return the "bare-metal" permutation (unwrap). For internal use only.

Access to wrapped permutation object. For "bare-metal" permutations this needs to return the identical (i.e. `===) object.

AbstractPermutations.permtype — Method

permtype(g::AbstractPermutation)

Return the group-theoretic type of permutation g, i.e. the vector of lengths of cycles in the (disjoint) cycle decomposition of g.

The lengths are sorted in decreasing order and cycles of length 1 are omitted. permtype(g) fully determines the conjugacy class of g in the full symmetric group.

Base.:^ — Method

^(i::Integer, σ::AbstractPermutation)

Return the image of i under the (permutation) action of σ.

We consider σ as a finite support permutation of ℕ, so by convention k^σ = k for all k > degree(σ).

Base.sign — Method

sign(g::AbstractPermutation)

Return the sign of a permutation.

sign represents the homomorphism from the permutation group to the unit group of ℤ whose kernel is the alternating group.

AbstractPermutations.@perm — Macro

@perm P cycles_string

Macro to parse cycles decomposition as a string into a permutation of type P.

Strings for the output of e.g. GAP could be copied directly into @perm, as long as they are not elided. Cycles of length 1 are not necessary, but can be included.

Examples:

julia> p = @perm Perm{UInt16} "(1,3)(2,4)"
(1,3)(2,4)

julia> typeof(p)
Perm{UInt16}

julia> q = @perm Perm "(1,3)(2,4)(3,5)(8)"
(1,5,3)(2,4)